Add .env variables to frontend. Minor action block css fix
