+++ react-frontend/src/helpers/Api.js
@@ -1,0 +1,37 @@
+
+class API {
+    static async deleteFile(fileDirToDelete) {
+        return await fetch(`${process.env.REACT_APP_API_URL}/delete/${encodeURIComponent(fileDirToDelete)}`, {
+            method: "DELETE"
+        })
+    }
+
+    static changeDir(socket, isConnected, dir, prevDir) {
+        if (isConnected) {
+          const data = {
+              type: "cd",
+              directory: './' + dir.join('/'),
+              prevDir: './' + prevDir.join('/')
+          }
+          socket.send(JSON.stringify(data));
+        }
+    }
+
+    static async moveFile () {
+
+    }
+
+    static async renameFile () {
+
+    }
+
+    static submitFile(reqBody) {
+        return fetch(`${process.env.REACT_APP_API_URL}/upload`, {
+            method: "POST",
+            body: reqBody
+        })
+    }
+
+}
+
+export default API
